+++ MaterialDesignExtensions/Converters/EmptyEnumerableToBoolConverter.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Data;
+
+namespace MaterialDesignExtensions.Converters
+{
+    public class EmptyEnumerableToBoolConverter : IValueConverter
+    {
+        public bool EmptyValue { get; set; }
+
+        public EmptyEnumerableToBoolConverter() { }
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is IEnumerable enumerable)
+            {
+                return enumerable.GetEnumerator().MoveNext() ? !EmptyValue : EmptyValue;
+            }
+            else
+            {
+                return EmptyValue;
+            }
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Binding.DoNothing;
+        }
+    }
+}